Move all Config and Subscription ServerWindow functionality into its specific files Use partial class and add two Files: - ServerWindowConfig - ServerWindowSubscription
