+++ InvisibleMan-XRay/Windows/ServerWindow/ServerWindow.xaml.cs
@@ -1,47 +1,20 @@
 using System;
-using System.Linq;
 using System.Windows;
 using System.Windows.Controls;
 using System.Collections.Generic;
-using Microsoft.Win32;
 
 namespace InvisibleManXRay
 {
     using Models;
     using Values;
     using Services;
-    using Services.Analytics.ServerWindow;
-    using Utilities;
 
     public partial class ServerWindow : Window
     {
-        private enum ImportingType { FILE, LINK }
-        private enum SubscriptionOperation { CREATE, EDIT }
-
-        private string configPath = null;
-        private string groupPath = null;
-        private ImportingType importingType;
-        private SubscriptionOperation subscriptionOperation;
-
-        private List<Components.Config> generalConfigComponents;
-        private List<Components.Config> subscriptionConfigComponents;
-
         private Func<string> getCurrentConfigPath;
         private Func<bool> isCurrentPathEqualRootConfigPath;
-        private Func<List<Config>> getAllGeneralConfigs;
-        private Func<string, List<Config>> getAllSubscriptionConfigs;
-        private Func<List<Subscription>> getAllGroups;
-        private Func<string, Status> convertLinkToConfig;
-        private Func<string, string, Status> convertLinkToSubscription;
-        private Func<string, Status> loadConfig;
         private Func<string, int> testConnection;
         private Func<string> getLogPath;
-        private Action<string> onCopyConfig;
-        private Action<string, string> onCreateConfig;
-        private Action<string, string, string> onCreateSubscription;
-        private Action<Subscription> onDeleteSubscription;
-        private Action<GroupType, string> onDeleteConfig;
-        private Action<string> onUpdateConfig;
 
         private AnalyticsService AnalyticsService => ServiceLocator.Get<AnalyticsService>();
 
@@ -49,20 +22,8 @@ namespace InvisibleManXRay
         {
             InitializeComponent();
             InitializeImportingGroups();
-            InitializeConfigComponents();
-
-            void InitializeImportingGroups()
-            {
-                SetActiveFileImportingGroup(true);
-                SetActiveLinkImportingGroup(false);
-                SetImportingType(ImportingType.FILE);
-            }
-
-            void InitializeConfigComponents()
-            {
-                generalConfigComponents = new List<Components.Config>();
-                subscriptionConfigComponents = new List<Components.Config>();
-            }
+            InitializeGeneralConfigComponents();
+            InitializeSubscriptionConfigComponents();
         }
 
         public void Setup(
@@ -108,11 +69,6 @@ namespace InvisibleManXRay
             LoadConfigsLists();
             ShowServersPanel();
             HandleShowingActiveTab(); 
-
-            void InitializeGroupPath()
-            {
-                groupPath = getCurrentConfigPath.Invoke();
-            } 
             
             void LoadConfigsLists()
             {
@@ -131,324 +87,9 @@ namespace InvisibleManXRay
             }
         }
 
-        private void OnConfigTabClick(object sender, RoutedEventArgs e)
-        {
-            EnableAllTabs();
-            HideAllPanels();
-
-            SetEnableConfigTabButton(false);
-            SetActiveConfigPanel(true);
-        }
-
-        private void OnSubscriptionTabClick(object sender, RoutedEventArgs e)
-        {
-            EnableAllTabs();
-            HideAllPanels();
-
-            SetEnableSubscriptionTabButton(false);
-            SetActiveSubscriptionPanel(true);
-        }
-
-        private void OnSubscriptionComboBoxSelectionChanged(object sender, RoutedEventArgs e)
-        {
-            if(comboBoxSubscription.SelectedValue == null)
-                return;
-
-            groupPath = ((Subscription)comboBoxSubscription.SelectedValue).Directory.FullName;
-            LoadConfigsList(GroupType.SUBSCRIPTION);
-            SelectConfig(getCurrentConfigPath.Invoke());
-        }
-
-        private void OnDeleteSubscriptionButtonClick(object sender, RoutedEventArgs e)
-        {
-            AnalyticsService.SendEvent(new SubDeleteButtonClickedEvent());
-
-            if(comboBoxSubscription.SelectedValue == null)
-                return;
-            
-            string remarks = comboBoxSubscription.Text;
-            Subscription subscription = (Subscription)comboBoxSubscription.SelectedValue;
-
-            MessageBoxResult result = MessageBox.Show(
-                this,
-                string.Format(Message.DELETE_CONFIRMATION, remarks),
-                Caption.INFO,
-                MessageBoxButton.YesNo,
-                MessageBoxImage.Question
-            );
-
-            if (result == MessageBoxResult.Yes)
-                DeleteSubscription(subscription);
-        }
-
-        private void OnEditSubscriptionButtonClick(object sender, RoutedEventArgs e)
-        {
-            AnalyticsService.SendEvent(new SubEditButtonClickedEvent());
-
-            if (comboBoxSubscription.SelectedValue == null)
-                return;
-
-            ShowEditSubscriptionServerPanel();
-        }
-
-        private void OnUpdateSubscriptionButtonClick(object sender, RoutedEventArgs e)
-        {
-            AnalyticsService.SendEvent(new SubUpdateButtonClickedEvent());
-            
-            InitializeTextBoxFields();
-            UpdateSubscription();
-
-            void InitializeTextBoxFields()
-            {
-                textBoxSubscriptionRemarks.Text = comboBoxSubscription.Text;
-                textBoxSubscriptionLink.Text = ((Subscription)comboBoxSubscription.SelectedValue).Url;
-            }
-
-            void UpdateSubscription()
-            {
-                EditSubscription(
-                    subscription: (Subscription)comboBoxSubscription.SelectedValue,
-                    remarks: textBoxSubscriptionRemarks.Text,
-                    link: textBoxSubscriptionLink.Text
-                );
-            }
-        }
-
-        private void OnAddConfigButtonClick(object sender, RoutedEventArgs e)
-        {
-            ShowAddConfigsServerPanel();
-            AnalyticsService.SendEvent(new AddConfigButtonClickedEvent());
-        }
-
-        private void OnAddSubscriptionButtonClick(object sender, RoutedEventArgs e)
-        {
-            ShowAddSubscriptionsServerPanel();
-            AnalyticsService.SendEvent(new AddSubButtonClickedEvent());
-        }
-
         private void OnCancelButtonClick(object sender, RoutedEventArgs e)
         {
             ShowServersPanel();
-        }
-
-        private void OnFileRadioButtonClick(object sender, RoutedEventArgs e)
-        {
-            SetActiveFileImportingGroup(true);
-            SetActiveLinkImportingGroup(false);
-            SetImportingType(ImportingType.FILE);
-        }
-
-        private void OnLinkRadioButtonClick(object sender, RoutedEventArgs e)
-        {
-            SetActiveFileImportingGroup(false);
-            SetActiveLinkImportingGroup(true);
-            SetImportingType(ImportingType.LINK);
-        }
-       
-        private void OnChooseFileButtonClick(object sender, RoutedEventArgs e)
-        {
-            OpenFileDialog fileDialog = InitializeFileDialog();
-            bool? result = fileDialog.ShowDialog();
-
-            if (!IsFileSelected())
-                return;
-
-            textBlockFileName.Text = GetFileName();
-            configPath = GetFilePath();
-
-            OpenFileDialog InitializeFileDialog()
-            {
-                OpenFileDialog fileDialog = new OpenFileDialog();
-                fileDialog.Title = "Add configuration";
-                fileDialog.Filter = "Config files|*.json;*.toml;*.yaml;*.yml|All files|*.*";
-
-                return fileDialog;
-            }
-
-            bool IsFileSelected() => result == true;
-
-            string GetFileName() => System.IO.Path.GetFileName(fileDialog.FileName);
-
-            string GetFilePath() => fileDialog.FileName;
-        }
-
-        private void OnImportConfigButtonClick(object sender, RoutedEventArgs e)
-        {
-            if (IsFileImporting())
-            {
-                HandleImportingConfigFromFile();
-                AnalyticsService.SendEvent(new ConfigFromFileImportedEvent());
-            }
-            else
-            {
-                HandleImportingConfigFromLink();
-                AnalyticsService.SendEvent(new ConfigFromLinkImportedEvent());
-            }
-
-            bool IsFileImporting() => importingType == ImportingType.FILE;
-
-            void HandleImportingConfigFromFile()
-            {
-                if (!IsFileSelected())
-                {
-                    MessageBox.Show(
-                        this,
-                        Values.Message.NO_CONFIG_FILE_SELECTED, 
-                        Values.Caption.WARNING, 
-                        MessageBoxButton.OK, 
-                        MessageBoxImage.Warning
-                    );
-                    return;
-                }
-
-                SetActiveLoadingPanel(true);
-                TryAddConfig(ConfigType.FILE);
-
-                bool IsFileSelected() => !string.IsNullOrEmpty(configPath);
-            }
-
-            void HandleImportingConfigFromLink()
-            {
-                if (!IsLinkEntered())
-                {
-                    MessageBox.Show(
-                        this,
-                        Values.Message.NO_CONFIG_LINK_ENTERED, 
-                        Values.Caption.WARNING, 
-                        MessageBoxButton.OK, 
-                        MessageBoxImage.Warning
-                    );
-                    return;
-                }
-
-                SetActiveLoadingPanel(true);
-                TryAddConfig(ConfigType.URL);
-
-                bool IsLinkEntered() => !string.IsNullOrEmpty(textBoxConfigLink.Text);
-            }
-
-            void TryAddConfig(ConfigType type)
-            {
-                Status configStatus;
-
-                if (type == ConfigType.FILE)
-                {
-                    configStatus = loadConfig.Invoke(configPath);
-                    if (configStatus.Code == Code.ERROR)
-                    {
-                        HandleError();
-                        SetActiveLoadingPanel(false);
-                        return;
-                    }
-
-                    onCopyConfig.Invoke(configPath);
-                }
-                else
-                {
-                    configStatus = convertLinkToConfig.Invoke(textBoxConfigLink.Text);
-                    if (configStatus.Code == Code.ERROR)
-                    {
-                        HandleError();
-                        SetActiveLoadingPanel(false);
-                        return;
-                    }
-
-                    string[] config = GetConfig();
-                    onCreateConfig.Invoke(GetConfigRemark(), GetConfigData());
-
-                    string[] GetConfig() => (string[])configStatus.Content;
-
-                    string GetConfigRemark() => config[0];
-
-                    string GetConfigData() => config[1];
-                }
-                
-                groupPath = GetLastConfigPath(GroupType.GENERAL);
-                onUpdateConfig.Invoke(GetLastConfigPath(GroupType.GENERAL));
-                SetActiveLoadingPanel(false);
-                LoadConfigsList(GroupType.GENERAL);
-                ClearConfigPath();
-                ClearConfigLink();
-                ShowServersPanel();
-
-                void HandleError()
-                {
-                    switch (configStatus.SubCode)
-                    {
-                        case SubCode.NO_CONFIG:
-                        case SubCode.UNSUPPORTED_LINK:
-                            HandleWarningMessage(configStatus.Content.ToString());
-                            break;
-                        case SubCode.INVALID_CONFIG:
-                            HandleErrorMessage(configStatus.Content.ToString());
-                            break;
-                        default:
-                            return;
-                    }
-                }
-            }
-        }
-
-        private void OnImportSubscriptionButtonClick(object sender, RoutedEventArgs e)
-        {
-            if (subscriptionOperation == SubscriptionOperation.CREATE)
-            {
-                AnalyticsService.SendEvent(new SubFromLinkImportedEvent());
-                AddSubscription();
-            }
-            else if (subscriptionOperation == SubscriptionOperation.EDIT)
-            {
-                AnalyticsService.SendEvent(new SubFromLinkEditedEvent());
-                EditSubscription(
-                    subscription: (Subscription)comboBoxSubscription.SelectedValue,
-                    remarks: textBoxSubscriptionRemarks.Text,
-                    link: textBoxSubscriptionLink.Text
-                );
-            }
-        }
-
-        private void ShowAddConfigsServerPanel()
-        {
-            panelServers.Visibility = Visibility.Hidden;
-            panelAddConfigs.Visibility = Visibility.Visible;
-        }
-
-        private void ShowAddSubscriptionsServerPanel()
-        {
-            subscriptionOperation = SubscriptionOperation.CREATE;
-            ShowSubscriptionPanel();
-            InitializeTextBoxFields();
-
-            void ShowSubscriptionPanel()
-            {
-                panelServers.Visibility = Visibility.Hidden;
-                panelAddSubscriptions.Visibility = Visibility.Visible;
-            }
-
-            void InitializeTextBoxFields()
-            {
-                textBoxSubscriptionRemarks.Text = "";
-                textBoxSubscriptionLink.Text = "";
-            }
-        }
-
-        private void ShowEditSubscriptionServerPanel()
-        {
-            subscriptionOperation = SubscriptionOperation.EDIT;
-            ShowSubscriptionPanel();
-            FetchTextBoxFields();
-
-            void ShowSubscriptionPanel()
-            {
-                panelServers.Visibility = Visibility.Hidden;
-                panelAddSubscriptions.Visibility = Visibility.Visible;
-            }
-
-            void FetchTextBoxFields()
-            {
-                textBoxSubscriptionRemarks.Text = comboBoxSubscription.Text;
-                textBoxSubscriptionLink.Text = ((Subscription)comboBoxSubscription.SelectedValue).Url;
-            }
         }
 
         private void ShowServersPanel()
@@ -459,123 +100,44 @@ namespace InvisibleManXRay
             SelectConfig(getCurrentConfigPath.Invoke());
         }
 
-        private void SetActiveFileImportingGroup(bool isActive)
-        {
-            ClearConfigLink();
-            buttonConfigFile.IsEnabled = isActive;
-        }
-
-        private void SetActiveLinkImportingGroup(bool isActive)
-        {
-            ClearConfigPath();
-            textBoxConfigLink.IsEnabled = isActive;
-        }
-
         private void SetImportingType(ImportingType type) => importingType = type;
-
-        private void ClearConfigPath()
-        {
-            configPath = null;
-            textBlockFileName.Text = "No file chosen...";
-        }
-
-        private void ClearConfigLink()
-        {
-            textBoxConfigLink.Text = null;
-        }
-
-        private void ClearSubscriptionRemarks()
-        {
-            textBoxSubscriptionRemarks.Text = null;
-        }
-
-        private void ClearSubscriptionPath()
-        {
-            textBoxSubscriptionLink.Text = null;
-        }
-
-        private void LoadGroupsList()
-        {
-            Dictionary<Subscription, string> groups;
-            InitializeGroups();
-            SelectCurrentGroup();
-
-            void InitializeGroups()
-            {
-                groups = new Dictionary<Subscription, string>();
-                foreach(Subscription group in getAllGroups.Invoke())
-                    groups.Add(group, group.Directory.Name);
-                
-                comboBoxSubscription.ItemsSource = groups;
-            }
-
-            void SelectCurrentGroup()
-            {
-                if (!IsAnyGroupExists())
-                    return;
-
-                KeyValuePair<Subscription, string> currentGroup = groups.FirstOrDefault(
-                    group => group.Key.Directory.FullName == FileUtility.GetDirectory(groupPath)
-                );
-
-                if (!IsCurrentGroupExists())
-                    currentGroup = groups.Last();
-                
-                comboBoxSubscription.SelectedValue = currentGroup.Key;
-
-                bool IsCurrentGroupExists() => currentGroup.Key != null && currentGroup.Value != null;
-            }
-
-            bool IsAnyGroupExists() => groups.Count > 0;
-        }
 
         private void LoadConfigsList(GroupType group)
         {
-            List<Components.Config> configComponents;
-            List<Config> configs;
-            List<Subscription> groups;
-            StackPanel parent;
-            TextBlock textNoServer;
-
             if (group == GroupType.GENERAL)
-            {
-                generalConfigComponents = new List<Components.Config>();
-                configs = getAllGeneralConfigs.Invoke();
-                groups = new List<Subscription>();
-                parent = listConfigs;
-                textNoServer = textNoConfig;
-                configComponents = generalConfigComponents;
-            }
+                LoadGeneralConfigsList();
             else
-            {
-                subscriptionConfigComponents = new List<Components.Config>();
-                configs = getAllSubscriptionConfigs.Invoke(groupPath);
-                groups = getAllGroups.Invoke();
-                parent = listSubscriptions;
-                textNoServer = textNoSubscription;
-                configComponents = subscriptionConfigComponents;
-                HandleShowingSubscriptionControlPanel();
-                
-                void HandleShowingSubscriptionControlPanel()
-                {
-                    if (groups.Count > 0)
-                        panelSubscriptionControl.Visibility = Visibility.Visible;
-                    else
-                        panelSubscriptionControl.Visibility = Visibility.Collapsed;
-                }
-            }
-            
-            ClearConfigsList(parent);
-            HandleShowingNoServerExistsHint(configs, groups, textNoServer);
+                LoadSubscriptionConfigsList();
+        }
 
+        private void ClearConfigsList(StackPanel list)
+        {
+            list.Children.Clear();
+        }
+
+        private void HandleShowingNoServerExistsHint(
+            List<Config> configs, 
+            List<Subscription> groups, 
+            TextBlock textNoServer
+        )
+        {
+            if (groups.Count == 0 && configs.Count == 0)
+                textNoServer.Visibility = Visibility.Visible;
+            else
+                textNoServer.Visibility = Visibility.Collapsed;
+        }
+
+        private void AddAllConfigsToList(
+            List<Config> configs,
+            List<Components.Config> configComponents,
+            StackPanel parent
+        )
+        {
             foreach (Config config in configs)
             {
                 Components.Config configComponent = CreateConfigComponent(config);
                 AddConfigToList(configComponent, configComponents, parent);
             }
-
-            if (IsAnyConfigExists(configs))
-                AddConfigHintAtTheEndOfList(parent);
 
             Components.Config CreateConfigComponent(Config config)
             {
@@ -631,59 +193,24 @@ namespace InvisibleManXRay
                 configComponentsList.Add(config);
                 parent.Children.Add(config);
             }
+        }
 
-            void ClearConfigsList(StackPanel list)
-            {
-                list.Children.Clear();
-            }
+        private bool IsAnyConfigExists(List<Config> configs)
+        {
+            return configs != null && configs.Count > 0;
+        }
 
-            void AddConfigHintAtTheEndOfList(StackPanel list)
-            {
-                list.Children.Add(new Components.ConfigHint());
-            }
-
-            bool IsAnyConfigExists(List<Config> configs)
-            {
-                return configs != null && configs.Count > 0;
-            }
-
-            void HandleShowingNoServerExistsHint(
-                List<Config> configs, 
-                List<Subscription> groups, 
-                TextBlock textNoServer
-            )
-            {
-                if (groups.Count == 0 && configs.Count == 0)
-                    textNoServer.Visibility = Visibility.Visible;
-                else
-                    textNoServer.Visibility = Visibility.Collapsed;
-            }
+        private void AddConfigHintAtTheEndOfList(StackPanel list)
+        {
+            list.Children.Add(new Components.ConfigHint());
         }
 
         private string GetLastConfigPath(GroupType group)
         {
-            Config lastConfig;
             if (group == GroupType.GENERAL)
-            {
-                lastConfig = getAllGeneralConfigs.Invoke().LastOrDefault();
-
-                if (!IsConfigExists())
-                    lastConfig = getAllSubscriptionConfigs.Invoke(groupPath).LastOrDefault();
-            }
+                return GetLastGeneralConfigPath();
             else
-            {
-                lastConfig = getAllSubscriptionConfigs.Invoke(groupPath).LastOrDefault();
-
-                if (!IsConfigExists())
-                    lastConfig = getAllGeneralConfigs.Invoke().LastOrDefault();
-            }
-                
-            if(!IsConfigExists())
-                return null;
-
-            return lastConfig.Path;
-
-            bool IsConfigExists() => lastConfig != null;
+                return GetLastSubscriptionConfigPath();
         } 
 
         private void SelectConfig(string path)
@@ -728,95 +255,6 @@ namespace InvisibleManXRay
             } 
         }
 
-        private void AddSubscription()
-        {
-            if (!IsRemarksEntered())
-            {
-                MessageBox.Show(
-                    this,
-                    Values.Message.NO_SUBSCRIPTION_REMARKS_ENTERED, 
-                    Values.Caption.WARNING, 
-                    MessageBoxButton.OK, 
-                    MessageBoxImage.Warning
-                );
-                return;
-            }
-            else if (!IsLinkEntered())
-            {
-                MessageBox.Show(
-                    this,
-                    Values.Message.NO_SUBSCRIPTION_LINK_ENTERED, 
-                    Values.Caption.WARNING, 
-                    MessageBoxButton.OK, 
-                    MessageBoxImage.Warning
-                );
-                return;
-            }
-
-            SetActiveLoadingPanel(true);
-            TryAddSubscription();
-
-            bool IsRemarksEntered() => !string.IsNullOrEmpty(textBoxSubscriptionRemarks.Text);
-
-            bool IsLinkEntered() => !string.IsNullOrEmpty(textBoxSubscriptionLink.Text);
-
-            void TryAddSubscription()
-            {
-                Status subscriptionStatus;
-
-                subscriptionStatus = convertLinkToSubscription.Invoke(
-                    textBoxSubscriptionRemarks.Text, 
-                    textBoxSubscriptionLink.Text
-                );
-
-                if (subscriptionStatus.Code == Code.ERROR)
-                {
-                    HandleError();
-                    SetActiveLoadingPanel(false);
-                    return;
-                }
-
-                string[] subscription = GetSubscription();
-                groupPath = "";
-                onCreateSubscription.Invoke(
-                    GetSubscriptionRemark(), 
-                    GetSubscriptionUrl(), 
-                    GetSubscriptionData()
-                );
-                onUpdateConfig.Invoke(GetLastConfigPath(GroupType.SUBSCRIPTION));
-                SetActiveLoadingPanel(false);
-                LoadGroupsList();
-                LoadConfigsList(GroupType.SUBSCRIPTION);
-                ClearSubscriptionRemarks();
-                ClearSubscriptionPath();
-                ShowServersPanel();
-
-                string[] GetSubscription() => (string[])subscriptionStatus.Content;
-
-                string GetSubscriptionUrl() => textBoxSubscriptionLink.Text;
-
-                string GetSubscriptionRemark() => subscription[0];
-
-                string GetSubscriptionData() => subscription[1];
-
-                void HandleError()
-                {
-                    switch (subscriptionStatus.SubCode)
-                    {
-                        case SubCode.NO_CONFIG:
-                        case SubCode.UNSUPPORTED_LINK:
-                            HandleWarningMessage(subscriptionStatus.Content.ToString());
-                            break;
-                        case SubCode.INVALID_CONFIG:
-                            HandleErrorMessage(subscriptionStatus.Content.ToString());
-                            break;
-                        default:
-                            return;
-                    }
-                }
-            }
-        }
-
         private void DeleteSubscription(Subscription subscription)
         {
             onDeleteSubscription.Invoke(subscription);
@@ -826,22 +264,7 @@ namespace InvisibleManXRay
             SelectConfig(getCurrentConfigPath.Invoke());
         }
 
-        private void EditSubscription(Subscription subscription, string remarks, string link)
-        {
-            string oldRemarks = comboBoxSubscription.Text;
-
-            AddSubscription();
-            if (!HasSameRemarks())
-                DeleteSubscription(subscription);
-
-            bool HasSameRemarks() => oldRemarks == comboBoxSubscription.Text;
-        }
-
         private void SetActiveLoadingPanel(bool isActive) => SetActivePanel(panelLoading, isActive);
-
-        private void SetActiveConfigPanel(bool isActive) => SetActivePanel(panelConfig, isActive);
-
-        private void SetActiveSubscriptionPanel(bool isActive) => SetActivePanel(panelSubscription, isActive);
 
         private void SetActivePanel(Panel panel, bool isActive)
         {
@@ -853,10 +276,6 @@ namespace InvisibleManXRay
             SetActiveConfigPanel(false);
             SetActiveSubscriptionPanel(false);
         }
-
-        private void SetEnableConfigTabButton(bool isEnabled) => SetEnableButton(buttonConfigTab, isEnabled);
-
-        private void SetEnableSubscriptionTabButton(bool isEnabled) => SetEnableButton(buttonSubscriptionTab, isEnabled);
 
         private void SetEnableButton(Button button, bool isEnabled)
         {
